+++ src/main/java/org/erachain/gui/items/persons/Persons_Favorite_TableModel.java
@@ -172,13 +172,15 @@ public class Persons_Favorite_TableModel extends TableModelCls<Tuple2<String, St
 
     public void removeObservers() {
 
-        Controller.getInstance().wallet.database.getPersonFavoritesSet().deleteObserver(this);
+        if (Controller.getInstance().doesWalletDatabaseExists())
+            Controller.getInstance().wallet.database.getPersonFavoritesSet().deleteObserver(this);
 
     }
 
     public void addObservers() {
 
-        Controller.getInstance().wallet.database.getPersonFavoritesSet().addObserver(this);
+        if (Controller.getInstance().doesWalletDatabaseExists())
+            Controller.getInstance().wallet.database.getPersonFavoritesSet().addObserver(this);
     }
